Change YamlElixir calls from `read_from_file/2` to `read_from_file!/2`
Disappointedly but unsurprisingly the behavior changed in a breaky way
with that big version jump from `1.3.0` to `2.1.0`.

`YamlElixir.read_from_file/2` is now returning a tuple of a status
code, then result. To get just the result we have to bang it.

All I really wanted to do with that reckless but hopefully harmless
deps upgrade was fix the credo issue.

Hadn’t run tests before the version upgrade commit (tsk). Banging the
yaml reads gets the tests passing again.